Base as of 11/25/25
Everything should be working now, it should be only letting valid inputs through and not throwing any tracebacks. (Yay!). Also all my sources are cited and docstrings have been written.

The only this that would be missing is adding the certificate at the top (but that's easy to copy and paste in at the end, and would need the assignment name/due date which we don't have yet), so not something to worry about rn!)